backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart
python-dotenv
google-generativeai>=0.3.2
firebase-admin>=6.4.0
pdfplumber>=0.10.4
requests>=2.31.0
pydantic>=2.6.0
httpx>=0.26.0
tiktoken>=0.5.2
pypdf>=4.0.1
openai>=1.0.0
langchain-groq
langchain
razorpay>=1.4.0
python-jobspy

# AutoApply dependencies
groq>=0.9.0
pydantic-settings>=2.4.0
sqlalchemy>=2.0.35
aiosqlite>=0.20.0
apscheduler>=3.10.4
playwright>=1.46.0
beautifulsoup4>=4.12.3
lxml>=5.3.0
aiohttp>=3.10.5
pymupdf>=1.24.10
reportlab>=4.2.2
jinja2>=3.1.4
aiosmtplib>=3.0.1
pandas>=2.2.2
pyyaml>=6.0.2
loguru>=0.7.2
tenacity>=9.0.0
